graph is based on btrfs uuids
the btrfs uuids are combined with the buildstep instructions. the
buildsteps are now held in /var/lib/conair/layers. since images are
snapshots of the last buildstep (layer) the parent uuid is taken if you
derive from that image. layer are not deleted yet so expect many
orphaned layers if you play around a lot.